+++ vantage-table/src/table/sets/insertable_value_set.rs
@@ -24,6 +24,7 @@ where
         let erased = self.as_entity_erased();
         let mut record = record.clone();
         run_before(self.before_insert_hooks(), &mut record, erased).await?;
+        self.strip_imported_columns(&mut record);
         enforce_invariants(&mut record, self.invariants())?;
         let id = self
             .data_source()